+++ ui/elements/options/numberoption.gd
@@ -17,21 +17,21 @@ func set_label(text : String) -> void:
 func set_option_name(name : String) -> void:
 	option_name = name
 
-func get_value() -> int:
+func get_value() -> float:
 	return $BaseOption/HBoxContainer/SpinBox.value
 
-func set_value(value : int) -> void:
+func set_value(value : float) -> void:
 	$BaseOption/HBoxContainer/SpinBox.value = value
 
 func set_widget_node(path : Node) -> void:
 	widget_node = path
 
 # Optional functions related to the specific control
-func set_spinbox_range(min_value : int, max_value : int, step : int) -> void:
+func set_spinbox_range(min_value : int, max_value : int, step : float) -> void:
 	$BaseOption/HBoxContainer/SpinBox.min_value = min_value
 	$BaseOption/HBoxContainer/SpinBox.max_value = max_value
 	$BaseOption/HBoxContainer/SpinBox.step = step
 
-func _on_SpinBox_value_changed(value : int) -> void:
+func _on_SpinBox_value_changed(value : float) -> void:
 	if widget_node:
 		widget_node.set_config_value(option_name, value)